requirements.txt
streamlit
pymongo
python-dotenv
requests
voyageai
langchain-core
pydantic
pandas
sentence-transformers
